refactor(agent): keep deployment CLI below size ratchet
